fix: 상품 관련 API 구현체를 인터페이스로 변경함 #14
